requirements.txt
pyyaml
jsonschema
openapi-spec-validator
rich          # pretty CLI output
networkx      # for dependency graph
pytest